+++ flutes/multiproc.py
@@ -607,11 +607,15 @@ class WriteEvent(NamedTuple):
     message: str
 
 
+class CloseEvent(NamedTuple):
+    worker_id: Optional[int]
+
+
 class QuitEvent(NamedTuple):
     pass
 
 
-Event = Union[NewEvent, UpdateEvent, WriteEvent, QuitEvent]
+Event = Union[NewEvent, UpdateEvent, WriteEvent, CloseEvent, QuitEvent]
 
 
 class ProgressBarManager:
@@ -651,58 +655,140 @@ class ProgressBarManager:
             for idx, _ in enumerate(pool.imap_unordered(run_fn, data)):
                 flutes.log(f"Processed {idx + 1} arrays")
 
+    :param verbose: If ``False``, all progress bars are disabled. Defaults to ``True``.
     :param kwargs: Default arguments for the `tqdm <https://tqdm.github.io/>`_ progress bar initializer.
     """
 
     class Proxy:
+        r"""Proxy class for the progress bar manager. Subprocesses should communicate with the progress bar manager
+        through this class.
+        """
         def __init__(self, queue: 'mp.Queue[Event]'):
             self.queue = queue
 
         @overload
-        def new(self, iterable: Iterable[T], **kwargs) -> Iterator[T]:
+        def new(self, iterable: Iterable[T], update_frequency: Union[int, float] = 1, **kwargs) -> Iterator[T]:
             ...
 
         @overload
-        def new(self, iterable: Literal[None] = None, **kwargs) -> tqdm:
+        def new(self, iterable: Literal[None] = None, update_frequency: Union[int, float] = 1, **kwargs) -> tqdm:
             ...
 
-        def new(self, iterable=None, **kwargs):
-            r"""Construct a new progress bar."""
+        def new(self, iterable=None, update_frequency=1, **kwargs):
+            r"""Construct a new progress bar.
+            
+            :param iterable: The iterable to decorate with a progress bar. If ``None``, then updates must be manually
+                managed with calls to :meth:`update`.
+            :param update_frequency: How many iterations per update. This argument only takes effect if :attr:`iterable`
+                is not ``None``:
+
+                - If :attr:`update_frequency` is a ``float``, then the progress bar is updated whenever the iterable
+                  progresses over that percentage of elements. For instance, a value of ``0.01`` results in an update
+                  per 1% of progress. Requires a sized iterable (having a valid ``__len__``).
+                - If :attr:`update_frequency` is an ``int``, then the progress bar is updated whenever the iterable
+                  progresses over that many elements. For instance, a value of ``10`` results in an update per 10
+                  elements.
+            :param kwargs: Additional arguments for the `tqdm <https://tqdm.github.io/>`_ progress bar initializer.
+                These can override the default arguments set in the constructor of :class:`ProgressBarManager`.
+            :return: The wrapped iterable, or the proxy class itself.
+            """
+            length = None
+            ret_val = self
             if iterable is not None:
-                length = None
                 try:
                     length = len(iterable)
                     kwargs.update(total=length)
                 except TypeError:
                     pass
+                if isinstance(update_frequency, float):
+                    if length is None:
+                        raise ValueError("`iterable` must have valid length if `update_frequency` is float")
+                    if not (0.0 < update_frequency <= 1.0):
+                        raise ValueError("`update_frequency` must be within the range (0, 1]")
+                    ret_val = self._iter_per_percentage(iterable, length, update_frequency)
+                else:
+                    if not (0 < update_frequency):
+                        raise ValueError("`update_frequency` must be positive")
+                    ret_val = self._iter_per_elems(iterable, update_frequency)
             self.queue.put_nowait(NewEvent(get_worker_id(), kwargs))
-            if iterable is not None:
-                return self._iter(iterable)
-            return self
+            return ret_val
 
-        def _iter(self, iterable: Iterable[T]) -> Iterator[T]:
-            for x in iterable:
+        def _iter_per_elems(self, iterable: Iterable[T], update_frequency: int) -> Iterator[T]:
+            prev_index = -1
+            next_index = update_frequency - 1
+            idx = 0
+            for idx, x in enumerate(iterable):
                 yield x
-                self.update(1)
+                if idx == next_index:
+                    self.update(idx - prev_index)
+                    next_index += update_frequency
+                    prev_index = idx
+            # At the end, `idx == len(iterable) - 1`.
+            if idx > prev_index:
+                self.update(idx - prev_index)
+
+        def _iter_per_percentage(self, iterable: Iterable[T], length: int, update_frequency: float) -> Iterator[T]:
+            update_count = 0
+            prev_index = -1
+            next_index = max(0, int(update_frequency * length) - 1)
+            for idx, x in enumerate(iterable):
+                yield x
+                if idx == next_index:
+                    self.update(idx - prev_index)
+                    update_count += 1
+                    next_index = max(idx + 1, int(update_frequency * (update_count + 1) * length) - 1)
+                    prev_index = idx
+            if length > prev_index + 1:
+                self.update(length - prev_index - 1)
 
         def update(self, n: int = 0, *, postfix: Optional[Dict[str, Any]] = None) -> None:
-            # TODO: Add throttle to prevent sending messages too often.
+            r"""Update progress for the current progress bar.
+
+            :param n: Increment to add to the counter.
+            :param postfix: An optional dictionary containing additional stats displayed at the end of the progress bar.
+                See `tqdm.set_postfix <https://tqdm.github.io/docs/tqdm/#set_postfix>`_ for more details.
+            """
             self.queue.put_nowait(UpdateEvent(get_worker_id(), n, postfix))
 
         def write(self, message: str) -> None:
+            r"""Write a message to console without disrupting the progress bars.
+
+            :param message: The message to write.
+            """
             self.queue.put_nowait(WriteEvent(get_worker_id(), message))
+
+        def close(self) -> None:
+            r"""Close the current progress bar."""
+            self.queue.put_nowait(CloseEvent(get_worker_id()))
 
         # Methods to imitate a normal progress bar.
         def __enter__(self):
             return self
 
         def __exit__(self, exc_type, exc_val, exc_tb):
+            self.close()
+
+    class DummyProxy(Proxy):
+        def new(self, iterable=None, **kwargs):
+            if iterable is not None:
+                return iterable
+            return self
+
+        def update(self, n: int = 0, *, postfix: Optional[Dict[str, Any]] = None) -> None:
             pass
 
-        def close(self):
+        def write(self, message: str) -> None:
             pass
 
-    def __init__(self, **kwargs):
+        def close(self) -> None:
+            pass
+
+    def __init__(self, verbose: bool = True, **kwargs):
+        self.verbose = verbose
+        if not verbose:
+            self._proxy = self.DummyProxy(None)
+            return
+
         self.manager = mp.Manager()
         self.queue: 'mp.Queue[Event]' = self.manager.Queue(-1)
         self.progress_bars: Dict[Optional[int], tqdm] = {}
@@ -719,7 +805,15 @@ class ProgressBarManager:
 
     @property
     def proxy(self):
+        r"""Return the proxy class for the progress bar manager. Subprocesses should communicate with the manager
+        through the proxy class.
+        """
         return self._proxy
+
+    def _close_bar(self, worker_id: Optional[int]) -> None:
+        if worker_id in self.progress_bars:
+            self.progress_bars[worker_id].close()
+            del self.progress_bars[worker_id]
 
     def _run(self):
         from tqdm import tqdm
@@ -728,9 +822,7 @@ class ProgressBarManager:
                 event = self.queue.get()
                 if isinstance(event, NewEvent):
                     position = self.worker_id_map[event.worker_id]
-                    if event.worker_id in self.progress_bars:
-                        self.progress_bars[event.worker_id].close()
-                        del self.progress_bars[event.worker_id]
+                    self._close_bar(event.worker_id)
                     kwargs = {**self.bar_kwargs, **event.kwargs, "leave": False, "position": position}
                     bar = tqdm(**kwargs)
                     self.progress_bars[event.worker_id] = bar
@@ -742,6 +834,8 @@ class ProgressBarManager:
                     bar.update(event.n)
                 elif isinstance(event, WriteEvent):
                     tqdm.write(event.message)
+                elif isinstance(event, CloseEvent):
+                    self._close_bar(event.worker_id)
                 elif isinstance(event, QuitEvent):
                     break
                 else:
@@ -760,6 +854,9 @@ class ProgressBarManager:
         self.close()
 
     def close(self):
+        if not self.verbose:
+            return
+
         self.queue.put_nowait(QuitEvent())
         self.thread.join()
         for bar in self.progress_bars.values():